+++ scruter-nextjs/app/(routes)/(footerPages)/about/page.tsx
