+++ scruter-nextjs/app/(routes)/(footerPages)/contributors/page.tsx
