+++ scruter-nextjs/app/(routes)/(footerPages)/faq/page.tsx
