+++ scruter-nextjs/app/(routes)/(footerPages)/help/page.tsx
