+++ scruter-nextjs/app/(routes)/(footerPages)/policy/page.tsx
