+++ scruter-nextjs/app/(routes)/(footerPages)/support/page.tsx
