+++ scruter-nextjs/app/(routes)/(footerPages)/team/page.tsx
