+++ scruter-nextjs/app/(routes)/(footerPages)/terms/page.tsx
