+++ scruter-nextjs/app/(routes)/(mainNavPages)/for-sale/page.tsx
